split component into files
