+++ src/components/SearchModal/CurrencySearch.tsx
@@ -11,7 +11,6 @@ import styled from 'styled-components'
 
 import InfoHelper from 'components/InfoHelper'
 import { KS_SETTING_API } from 'constants/env'
-import { Z_INDEXS } from 'constants/styles'
 import { nativeOnChain } from 'constants/tokens'
 import { AllTokenType, useAllTokens, useToken } from 'hooks/Tokens'
 import useDebounce from 'hooks/useDebounce'
@@ -19,7 +18,6 @@ import { useOnClickOutside } from 'hooks/useOnClickOutside'
 import usePrevious from 'hooks/usePrevious'
 import useTheme from 'hooks/useTheme'
 import useToggle from 'hooks/useToggle'
-import { useBridgeState } from 'state/bridge/hooks'
 import { WrappedTokenInfo } from 'state/lists/wrappedTokenInfo'
 import { useRemoveUserAddedToken, useUserAddedTokens, useUserFavoriteTokens } from 'state/user/hooks'
 import { filterTokens } from 'utils/filtering'
@@ -30,7 +28,7 @@ import { filterTruthy, isAddress } from '../../utils'
 import Column from '../Column'
 import { RowBetween } from '../Row'
 import CommonBases from './CommonBases'
-import CurrencyList, { CurrencyListBridge } from './CurrencyList'
+import CurrencyList from './CurrencyList'
 import { useTokenComparator } from './sorting'
 import { PaddedColumn, SearchIcon, SearchInput, SearchWrapper, Separator } from './styleds'
 
@@ -39,7 +37,7 @@ enum Tab {
   Imported,
 }
 
-const ContentWrapper = styled(Column)`
+export const ContentWrapper = styled(Column)`
   width: 100%;
   flex: 1 1;
   position: relative;
@@ -136,7 +134,7 @@ const fetchTokens = async (
   }
 }
 
-const NoResult = () => {
+export const NoResult = () => {
   const theme = useTheme()
   return (
     <Column style={{ padding: '20px', height: '100%' }}>
@@ -526,160 +524,6 @@ export function CurrencySearch({
             hasMore={hasMoreToken}
           />
         </div>
-      ) : (
-        <NoResult />
-      )}
-    </ContentWrapper>
-  )
-}
-
-interface CurrencySearchBridgeProps {
-  isOpen: boolean
-  onDismiss: () => void
-  isOutput?: boolean
-  onCurrencySelect: (currency: WrappedTokenInfo) => void
-}
-
-export function CurrencySearchBridge({ isOutput, onCurrencySelect, onDismiss, isOpen }: CurrencySearchBridgeProps) {
-  const theme = useTheme()
-
-  const [searchQuery, setSearchQuery] = useState<string>('')
-  const debouncedQuery = useDebounce(searchQuery, 200)
-
-  const [{ listTokenIn, listTokenOut }] = useBridgeState()
-
-  const fetchedTokens = isOutput ? listTokenOut : listTokenIn
-  const tokenComparator = useTokenComparator(false, true)
-
-  const isAddressSearch = isAddress(debouncedQuery)
-
-  const filteredTokens: WrappedTokenInfo[] = useMemo(() => {
-    if (isAddressSearch) {
-      const find = fetchedTokens.find(e => e?.address?.toLowerCase() === debouncedQuery.toLowerCase())
-      return find ? [find] : []
-    }
-    return filterTokens(fetchedTokens, debouncedQuery)
-  }, [isAddressSearch, fetchedTokens, debouncedQuery])
-
-  const visibleCurrencies: WrappedTokenInfo[] = useMemo(() => {
-    const sorted = filteredTokens.sort(tokenComparator)
-    const symbolMatch = debouncedQuery
-      .toLowerCase()
-      .split(/\s+/)
-      .filter(s => s.length > 0)
-    if (symbolMatch.length > 1) return sorted
-
-    return [
-      // sort any exact symbol matches first
-      ...sorted.filter(token => token.symbol?.toLowerCase() === symbolMatch[0]),
-      ...sorted.filter(token => token.symbol?.toLowerCase() !== symbolMatch[0]),
-    ]
-  }, [filteredTokens, debouncedQuery, tokenComparator])
-
-  const handleCurrencySelect = useCallback(
-    (currency: WrappedTokenInfo) => {
-      onCurrencySelect(currency)
-      onDismiss()
-    },
-    [onDismiss, onCurrencySelect],
-  )
-
-  // manage focus on modal show
-  const inputRef = useRef<HTMLInputElement>(null)
-
-  // clear the input on open
-  useEffect(() => {
-    if (isOpen) {
-      setSearchQuery('')
-      inputRef.current?.focus()
-    }
-  }, [isOpen])
-
-  const handleInput = useCallback((event: ChangeEvent<HTMLInputElement>) => {
-    const input = event.target.value
-    const checksumInput = isAddress(input)
-    setSearchQuery(checksumInput || input)
-  }, [])
-
-  const handleEnter = useCallback(
-    (e: KeyboardEvent<HTMLInputElement>) => {
-      if (
-        e.key === 'Enter' &&
-        visibleCurrencies.length > 0 &&
-        (visibleCurrencies[0].symbol?.toLowerCase() === searchQuery.trim().toLowerCase() ||
-          visibleCurrencies.length === 1)
-      ) {
-        handleCurrencySelect(visibleCurrencies[0])
-      }
-    },
-    [visibleCurrencies, handleCurrencySelect, searchQuery],
-  )
-
-  // menu ui
-  const [open, toggle] = useToggle(false)
-  const node = useRef<HTMLDivElement>()
-  useOnClickOutside(node, open ? toggle : undefined)
-
-  return (
-    <ContentWrapper>
-      <PaddedColumn gap="14px">
-        <RowBetween>
-          <Text fontWeight={500} fontSize={20} display="flex">
-            <Trans>Select a token</Trans>
-            {!isOutput && (
-              <InfoHelper
-                zIndexTooltip={Z_INDEXS.MODAL + 1}
-                size={16}
-                text={
-                  <Trans>You can select and transfer any token supported by Multichain from one chain to another</Trans>
-                }
-              />
-            )}
-          </Text>
-          <CloseIcon onClick={onDismiss} />
-        </RowBetween>
-
-        {!isOutput ? (
-          <SearchWrapper>
-            <SearchInput
-              type="text"
-              id="token-search-input"
-              placeholder={t`Search by token name, token symbol or address`}
-              value={searchQuery}
-              ref={inputRef}
-              onChange={handleInput}
-              onKeyDown={handleEnter}
-              autoComplete="off"
-            />
-            <SearchIcon size={18} color={theme.border} />
-          </SearchWrapper>
-        ) : (
-          <>
-            <Text fontSize={12} color={theme.subText} lineHeight="20px">
-              <Trans>
-                You can select from one of the token pools below. Different pools may have different liquidity and fees
-              </Trans>
-            </Text>
-            <Flex justifyContent="space-between" color={theme.subText} fontWeight={500} fontSize={15}>
-              <Text>
-                <Trans>Pool</Trans>
-              </Text>
-              <Text>
-                <Trans>Liquidity</Trans>
-              </Text>
-            </Flex>
-          </>
-        )}
-      </PaddedColumn>
-
-      <Separator />
-
-      {visibleCurrencies?.length ? (
-        <CurrencyListBridge
-          isOutput={isOutput}
-          currencies={visibleCurrencies}
-          onCurrencySelect={handleCurrencySelect}
-        />
       ) : (
         <NoResult />
       )}
